% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_model.R, R/disequilibrium_model.R,
%   R/diseq_basic.R, R/diseq_deterministic_adjustment.R, R/diseq_directional.R,
%   R/diseq_stochastic_adjustment.R, R/equilibrium_model.R
\docType{class}
\name{market_models}
\alias{market_models}
\alias{disequilibrium_model-class}
\alias{diseq_basic-class}
\alias{diseq_deterministic_adjustment-class}
\alias{diseq_directional-class}
\alias{diseq_stochastic_adjustment-class}
\alias{equilibrium_model-class}
\title{Market model classes}
\description{
\subsection{Basic disequilibrium model with unknown sample separation
(\code{diseq_basic})}{
The basic disequilibrium model consists of three equations. Two of them
are the demand and supply equations. In addition, the model replaces the
market clearing condition with the short side rule. The model is estimated
using full information maximum likelihood.

\deqn{D_{nt} = X_{d,nt}'\beta_{d} + u_{d,nt},}
\deqn{S_{nt} = X_{s,nt}'\beta_{s} + u_{s,nt},}
\deqn{Q_{nt} = \min\{D_{nt},S_{nt}\}.}
}

\subsection{Disequilibrium model with deterministic price dynamics
(\code{diseq_deterministic_adjustment})}{
The disequilibrium model with deterministic price adjustment consists
of four equations. The two market equations, the short side rule and price
evolution equation. The first two equations are stochastic. The price
equation is deterministic. The sample is separated based on the sign of
the price changes as in the \code{\linkS4class{diseq_directional}} model.
The model is estimated using full information maximum likelihood.

\deqn{D_{nt} = X_{d,nt}'\beta_{d} + P_{nt}\alpha_{d} + u_{d,nt},}
\deqn{S_{nt} = X_{s,nt}'\beta_{s} + P_{nt}\alpha_{s} + u_{s,nt},}
\deqn{Q_{nt} = \min\{D_{nt},S_{nt}\},}
\deqn{\Delta P_{nt} = \frac{1}{\gamma} \left( D_{nt} - S_{nt} \right).}
}

\subsection{Directional disequilibrium model with sample separation
(\code{diseq_directional})}{
The directional disequilibrium model consists of three equations and a
separation rule. The market is described by a linear demand, a linear supply
equation and the short side rule. The separation rule splits the sample
into states of excess supply and excess demand. If a price change is
positive at the time point of the observation, then the observation is
classified as being in an excess demand state. Otherwise, it is assumed
that it represents an excess supply state. The
model is estimated using full information maximum likelihood.

\deqn{D_{nt} = X_{d,nt}'\beta_{d} + u_{d,nt},}
\deqn{S_{nt} = X_{s,nt}'\beta_{s} + u_{s,nt},}
\deqn{Q_{nt} = \min\{D_{nt},S_{nt}\},}
\deqn{\Delta P_{nt} \ge 0 \Longrightarrow D_{nt} \ge S_{nt}.}
}

\subsection{Disequilibrium model with stochastic price dynamics
(\code{diseq_stochastic_adjustment})}{
The disequilibrium model with stochastic price adjustment is described
by a system of four equations. Three of of them form a stochastic linear system of
market equations equations coupled with a stochastic price evolution equation. The
fourth equation is the short side rule. In contrast to the deterministic counterpart,
the model does not impose any separation rule on the sample. It is estimated using
full information maximum likelihood.

\deqn{D_{nt} = X_{d,nt}'\beta_{d} + P_{nt}\alpha_{d} + u_{d,nt},}
\deqn{S_{nt} = X_{s,nt}'\beta_{s} + P_{nt}\alpha_{s} + u_{s,nt},}
\deqn{Q_{nt} = \min\{D_{nt},S_{nt}\},}
\deqn{\Delta P_{nt} =
    \frac{1}{\gamma} \left( D_{nt} - S_{nt} \right) +  X_{p,nt}'\beta_{p} + u_{p,nt}.}
}

\subsection{Equilibrium model (\code{equilibrium_model})}{
The equilibrium model consists of thee equations. The demand, the
supply and the market clearing equations. The model can be estimated using both full
information maximum likelihood and two-stage least squares.

\deqn{D_{nt} = X_{d,nt}'\beta_{d} + P_{nt}\alpha_{d} + u_{d,nt},}
\deqn{S_{nt} = X_{s,nt}'\beta_{s} + P_{nt}\alpha_{s} + u_{s,nt},}
\deqn{Q_{nt} = D_{nt} = S_{nt}.}

A necessary identification condition is that
there is at least one control that is exclusively part of the demand and one control
that is exclusively part of the supply equation. In the first stage of the two-stage
least square estimation, prices are regressed on remaining controls from both the
demand and supply equations. In the second stage, the demand and supply equation is
estimated using the fitted prices instead of the observed.
}
}
\section{Slots}{

\describe{
\item{\code{logger}}{Logger object.}

\item{\code{subject_columns}}{Column name for the subject identifier.}

\item{\code{time_column}}{Column name for the time point identifier.}

\item{\code{explanatory_columns}}{Vector of explanatory column names for all model's
equations.}

\item{\code{data_columns}}{Vector of model's data column names. This is the union of the
quantity, price and explanatory columns.}

\item{\code{columns}}{Vector of primary key and data column names for all model's equations.}

\item{\code{data}}{Model data frame.}

\item{\code{model_name}}{Model name string.}

\item{\code{market_type}}{Market type string.}

\item{\code{system}}{Model's system of equations.}
}}

\seealso{
\link{model_initialization}
}
