% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_fit.R
\name{plot,market_fit-method}
\alias{plot,market_fit-method}
\title{Plots the fitted model}
\usage{
\S4method{plot}{market_fit}(x, subject, time, show_scatter = TRUE, ...)
}
\arguments{
\item{x}{A model object.}

\item{subject}{A vector of subject identifiers to be used in the
visualization.}

\item{time}{A vector of time identifiers to be used in the visualization.}

\item{show_scatter}{Should the price-quantity scatter be plotted? By default
\code{TRUE}.}

\item{...}{Additional parameters to be used for styling the figure.
Specifically \code{xlab}, \code{ylab}, and \code{main} are currently
handled by the function.}
}
\value{
No return value, called for for side effects (visualization).
}
\description{
Displays a graphical illustration of the passed fitted model object. The
function creates a scatter plot of quantity-price pairs for the records
corresponding to the given subject and time identifiers. Then, it plots
the average fitted demand and supply quantities for the same data subset
letting prices vary between the minimum and maximum price
points observed in the data subset.
}
\details{
If the \code{subject} argument is missing, all subjects are used. If the
\code{time} argument is missing, all time points are used. The scatter
plot of the quantity-price data can be suppressed by setting
\code{show_scatter = FALSE}.
}
\examples{
\donttest{
# estimate a model using the houses dataset
fit <- diseq_deterministic_adjustment(
  HS | RM | ID | TREND ~
    RM + TREND + W + CSHS + L1RM + L2RM + MONTH |
      RM + TREND + W + L1RM + MA6DSF + MA3DHF + MONTH,
  fair_houses(),
  correlated_shocks = FALSE,
  estimation_options = list(control = list(maxit = 1e+6))
)

# show model's illustration plot
plot(fit)
}
}
